Adds a command to download Cake bootstrapper
This command installs the Cake's bootstrapper to the platform selected
by user. The bootstrapper is downloaded from the Cake repository and
saved in the root folder.

Unfortunately the VSCode doesn't have a UI control that allow multiple
item selection, therefore the user needs to run the command for each
platform they needs the bootstrapper.

To make the user's life a little bit easier, the platform on which it
is running VSCode will always be displayed on top of the list, that way
he can select the item just pressing `ENTER`