(GH-308) Fix usage of undefined Tools
The issue here was that an attempt was being made to install the debug
dependencies into the Tools folder before the cake.config file has been
fully downloaded, so it wasn't able to parse the ini file correctly.
Move the installation of the debug tools to after all the download have
happened.