(GH-432) Only download requested bootstrappers
When installing Cake to the current workspace, the option is provided to
download bootstrappers, however, there are now more than the normal two
bootstrappers.  As a result, when downloading all files, there is a
problem since the file names are the same.  To prevent this, prompt the
user to select what type of bootstrappers they want to install.