(#515) Automatically enable Omnisharp usage
Once downloaded, the Cake.Bakery.exe can be added to the Omnisharp
configuration file, so that Omnisharp knows where to find the binary,
and then once the file is updated, force a restart of the Omnisharp
Server so that everything just "works".

Here we are relying on the fact that any VSCode extension can call a
registered command from another extension, as long as the name is
known.  In this case, the `o.restart` command, which the Omnisharp
Extension registers.